Added metadataProvider to PsrLogger
